fixing nan issues in project 48

+++ project48/main.cpp
@@ -5,7 +5,7 @@ using namespace std;
 
 long double calculateSelfPowers(int size){
 	long double count = 0;
-	for(int i = 1; i <= size; i++){
+	for(long double i = 1; i <= size; i++){
 		count += pow(i,i);
 	}
 }
